backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
cffi==2.0.0
click==8.3.0
cloudpickle==3.1.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
ecdsa==0.19.1
fastapi==0.120.0
fonttools==4.60.1
greenlet==3.2.4
h11==0.16.0
idna==3.11
joblib==1.5.2
kiwisolver==1.4.9
llvmlite==0.45.1
matplotlib==3.10.7
numba==0.62.1
numpy==2.3.4
nvidia-nccl-cu12==2.28.7
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pillow==12.0.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
shap==0.49.1
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette==0.48.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
xgboost==3.1.1
